Update and rename Hierarchical.py to hierarchical.py
